backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.6
argostranslate==1.9.6
kokoro-onnx==0.2.1
numpy==1.26.4
python-multipart==0.0.20
